+++ skysail.domain/etc/changelog.md
@@ -1,7 +1,8 @@
 new in
 
-0.3.0 
+6.0.0 
 -----
 
- * new: GenericIdentifiable
+ * new: GenericIdentifiable in io.skysail.domain 0.3.0
+ * removed dependency on restlet in io.skysail.domain.html 2.0.0 (https://github.com/evandor/skysail/issues/10)